+++ src/App.svelte
@@ -1,10 +1,39 @@
+
 <script>
-	export let name;
+  import CV from './CV.svelte'
+  import HttpDemo from './HttpDemo.svelte'
+
+
+  function handleNewSkill(newSkill) {
+    const { detail: { skill, level } } = newSkill;
+    person.skills = [...person.skills, { name: skill, level }];
+  }
+
+  let person = {
+    name: 'Tankred',
+    skills: [
+      {
+        name: 'Svelte',
+        level: 5
+      },
+      {
+        name: 'Javascript',
+        level: 5
+      },
+      {
+        name: 'Security',
+        level: 5
+      }
+    ]
+  }
+
 </script>
 
 <main>
-	<h1>Hello {name}!</h1>
+	<CV person={person} on:newSkill={handleNewSkill} />
+	<h1>Hello {person.name}!</h1>
 	<p>Visit the <a href="https://svelte.dev/tutorial">Svelte tutorial</a> to learn how to build Svelte apps.</p>
+	<HttpDemo /> 
 </main>
 
 <style>