+++ CHANGELOG.md
@@ -6,6 +6,48 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [Released]
+
+## [0.3.0] - 2021-01-27
+### Changed
+- Build and publish live lab
+
+## [0.2.8] - 2021-01-26
+### Changed
+- Fix bug max int in skill level eq 5
+
+## [0.2.7] - 2021-01-22
+### Added
+- LICENSE.md
+
+## [0.2.6] - 2021-01-20
+### Changed
+- Broadcast skills 
+
+## [0.2.5] - 2021-01-19
+### Changed
+- Add a skill input form
+
+## [0.2.4] - 2021-01-18
+### Added
+- Add Http demo
+
+## [0.2.3] - 2021-01-15
+### Changed
+- Add IF ELSE conditionals
+
+## [0.2.2] - 2021-01-14
+### Changed
+- Add a for loop (list) 
+
+## [0.2.1] - 2021-01-13
+### Changed
+- Add props
+
+## [0.2.0] - 2021-01-12
+### Added
+- Add component CV 
+
 ## [0.1.4] - 2021-01-11
 ### Changed
 - Live reload: check
@@ -14,8 +56,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 npm run dev
 
 ```
-
-## [Released]
 
 ## [0.1.3] - 2021-01-08
 ### Added